// SPA/src/app/_core/models/transaction/cancel-sticker.ts
export interface CancelStickerViewModel {
  isSortScan: string;
  grade: string;
  wkshno: string;
  manNo: string;
  purNo: string;
  size: string;
  serial: number;
  qty: number;
  prt_Cnt: number;
  qrCodeID: string;
  qrCodeValue: string;
  empno: string;
  crUsr: string;
  prtDay: string;
  isChecked?: boolean;
}